Fix problem with the functions that return null when they are operated with other parameters
Fix problem with the functions that return null when they are operated with other parameters, either in sum, divisions, etc. or in comparisons.
For comparisons it is necessary to allow null parameters with AllowNullParameter
Modify comparison method since when comparing a parameter with null value and an integer, it failed.
Add tests to validate the changes made
Fix test that failed because the message of the exception was in Spanish and in the test was compared with a message in English